# doc/api.yaml
openapi: 3.0.0
paths:
...
  /user:
    {
      get:
        {
          operationId: UsersController_getAllUsers,
          parameters: [],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/User' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [users],
          security: [{ bearer: [] }],
        },
    }
  '/user/{id}':
    {
      get:
        {
          operationId: UsersController_getUserById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/User' } },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'User not found' },
            },
          tags: [users],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: UsersController_updateUser,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    {
                      schema:
                        { $ref: '#/components/schemas/UpdatePasswordDto' },
                    },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/User' } },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '404': { description: 'Track not found' },
            },
          tags: [users],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: UsersController_deleteUser,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'User not found' },
            },
          tags: [users],
          security: [{ bearer: [] }],
        },
    }
...
  '/favs/track/{id}':
    {
      post:
        {
          operationId: FavsController_addTrackToFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '201': { description: 'Successfully added to favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '422': { description: 'Track does not exist' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: FavsController_deleteTrackFromFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '204': { description: 'Successfully deleted from favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Track is not favourite' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
...
  /track:
    {
      get:
        {
          operationId: TracksController_getAllTracks,
          parameters: [],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Track' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
      post:
        {
          operationId: TracksController_createTrack,
          parameters: [],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/TrackDto' } },
                },
            },
          responses:
            {
              '201':
                {
                  description: 'Successfully created',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Track' } },
                    },
                },
              '400': { description: 'Invalid dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
    }
  /track/search:
    {
      get:
        {
          operationId: TracksController_getTracksByName,
          parameters:
            [
              {
                name: name,
                required: true,
                in: query,
                description: 'Name of the track',
                schema: { type: string },
              },
            ],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Track' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
    }
  '/track/{id}':
    {
      get:
        {
          operationId: TracksController_getTrackById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Track' } },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '404': { description: 'Track not found' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: TracksController_updateTrack,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/TrackDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Track' } },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Track not found' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: TracksController_deleteTrack,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album not found' },
            },
          tags: [tracks],
          security: [{ bearer: [] }],
        },
    }
...
info:
  title: 'Home Library'
  description: 'Home music library service'
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes: { bearer: { scheme: bearer, bearerFormat: JWT, type: http } }
  schemas:
...
      User:
        {
          type: object,
          properties:
            {
              id: { type: string },
              login: { type: string },
              email: { type: string },
              emailIsConfirmed: { type: boolean },
              version: { type: number },
              createdAt: { type: number },
              updatedAt: { type: number },
              role: { type: string },
            },
          required:
            [
              id,
              login,
              email,
              emailIsConfirmed,
              version,
              createdAt,
              updatedAt,
              role,
            ],
        },
      UpdatePasswordDto:
        {
          type: object,
          properties:
            { oldPassword: { type: string }, newPassword: { type: string } },
          required: [oldPassword, newPassword],
        },
      FindFavsDto:
        {
          type: object,
          properties: { userId: { type: string } },
          required: [userId],
        },
...
      Track:
        {
          type: object,
          properties:
            {
              id: { type: string },
              name: { type: string },
              duration: { type: number },
              artistId: { type: string },
              albumId: { type: string },
            },
          required: [id, name, duration, artistId, albumId],
        },
      TrackDto:
        {
          type: object,
          properties:
            {
              name: { type: string },
              artistId: { type: string },
              albumId: { type: string },
              duration: { type: number },
            },
          required: [name, duration],
        },
...